move blog schema
