+++ src/content/config.ts
@@ -1,6 +1,6 @@
 import { defineCollection, z, type SchemaContext } from "astro:content";
 import { docsSchema } from "@astrojs/starlight/schema";
-import { blogSchema } from "./blogSchema";
+import { blogSchema } from "../blogSchema";
 
 export const collections = {
   docs: defineCollection({
